Fix bug in life systmem Set DeathAnimation on life 0

+++ Assets/Components/EventSystem/EventSystem.cs
@@ -4,7 +4,7 @@ namespace Components.EventSystem
 {
     public static class Events
     {
-        public static Action<int> OnLifeCountChanged;
+        public static Action<float> OnLifeCountChanged;
         public static Action<float> UpdateLife;
     }
 }
